Add files via upload
What was difficult about this project was trying to find out how to make the text to speech say the number of times I clicked the button. One I found out how, there was one more difficult task which was trying to add the "Cookies ate" speech after the button was clicked, which I unfortunately was not able to do.